# backend/openapi.yaml
openapi: 3.1.0
info:
  title: "KAVACH \u2014 AI-Powered DevSecOps Platform"
  description: "KAVACH scans banking application repositories for vulnerabilities,\
    \ calculates a Banking Risk Score (BRS), maps findings to regulatory controls\
    \ (RBI / PCI-DSS / SWIFT), and generates AI-powered audit reports.\n\n## Pipeline\n\
    `GitHub webhook` \u2192 `Create Scan Job` \u2192 `Queue` \u2192 `Distributed workers`\
    \ \u2192 `Aggregation` \u2192 `Risk Engine` \u2192 `Compliance Engine` \u2192\
    \ `AI Explanation` \u2192 `Reports` \u2192 `Dashboard` \u2192 `Notifications`\
    \ \u2192 `Archive`\n\nA scan enters this pipeline either via `POST /scan/repository`\
    \ / `POST /scan` (an authenticated user submitting a URL or archive) or `POST\
    \ /webhooks/github` (an unauthenticated, HMAC-verified push event) \u2014 both\
    \ converge on the same queue \u2192 distributed-worker \u2192 aggregation path,\
    \ so everything downstream behaves identically regardless of how the scan was\
    \ triggered."
  contact:
    name: KAVACH Platform Team
    url: https://github.com/kavach-platform/kavach
  license:
    name: "Proprietary \u2014 All Rights Reserved"
  version: 1.0.0
paths:
...
  /api/v1/scan/repository:
    post:
      tags:
      - Scanning
      summary: Submit Repository
      description: "Submit a repository by URL: validate it, resolve (or reuse) its\n\
        Repository row, create a queued ScanJob, and return immediately. The\narchive\
        \ download and the 9-scanner fan-out run asynchronously \u2014 see\napp/tasks/scan_tasks.py.\
        \ Submitting the same URL again reuses the same\nRepository row (see app/services/scan_intake.py)\
        \ rather than creating\na duplicate."
      operationId: submit_repository_api_v1_scan_repository_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanJobSubmitRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanJobCreateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
...
  /api/v1/webhooks/github:
    post:
      tags:
      - Webhooks
      - Webhooks
      summary: GitHub push event receiver
      description: "Handles two GitHub event types:\n\n- `ping` \u2014 sent once when\
        \ the webhook is first configured, to let\n  GitHub confirm the endpoint is\
        \ reachable and signs correctly.\n  Acknowledged without creating a scan.\n\
        - `push` \u2014 the trigger. Resolves (or reuses) the pushed repository's\n\
        \  `Repository` row, creates a queued `ScanJob` for the pushed branch,\n \
        \ and dispatches it exactly like `POST /scan/repository` \u2014 the rest\n\
        \  of the pipeline (workers, aggregation, risk/compliance/AI, reports,\n \
        \ notifications, archive) runs identically regardless of how the scan\n  was\
        \ triggered.\n\nEvery other event type (`pull_request`, `issues`, `star`,\
        \ ...) is\nacknowledged with 200 and a `status: \"ignored\"` body rather than\n\
        erroring \u2014 GitHub retries/disables a webhook that responds with\nnon-2xx\
        \ too often, and subscribing to more events than intended on\nthe GitHub side\
        \ shouldn't be able to break delivery for the ones that\nmatter.\n\nResponds\
        \ fast (queues the job, does not wait for it): GitHub expects\na response\
        \ within 10 seconds and will consider the delivery failed\n(and may retry)\
        \ otherwise."
      operationId: github_webhook_api_v1_webhooks_github_post
      parameters:
      - name: x-hub-signature-256
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Hub-Signature-256
      - name: x-github-event
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Github-Event
      responses:
        '200':
          description: Whether a scan was queued, ignored, or the delivery was just
            a ping check
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookAckResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/auth/admin/users:
    get:
      tags:
      - "Auth \u2014 Admin"
      summary: List Users
      operationId: list_users_api_v1_auth_admin_users_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserRead'
                title: Response List Users Api V1 Auth Admin Users Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/admin/users/{user_id}/role:
    patch:
      tags:
      - "Auth \u2014 Admin"
      summary: Update User Role
      operationId: update_user_role_api_v1_auth_admin_users__user_id__role_patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/admin/users/{user_id}/active:
    patch:
      tags:
      - "Auth \u2014 Admin"
      summary: Update User Active Status
      operationId: update_user_active_status_api_v1_auth_admin_users__user_id__active_patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActiveStatusUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/audit-log:
    get:
      tags:
      - "Auth \u2014 Admin"
      summary: Query Audit Log
      description: 'Defaults to the last `audit_log_default_lookback_days` days (90
        by

        default) when `since` isn''t given, so a careless unbounded query can''t

        accidentally full-scan the entire table.'
      operationId: query_audit_log_api_v1_auth_audit_log_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: User Id
      - name: action
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Action
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: since
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Since
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ActiveStatusUpdateRequest:
      properties:
        is_active:
          type: boolean
          title: Is Active
      type: object
      required:
      - is_active
      title: ActiveStatusUpdateRequest
    AuditLogEntryResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: User Id
        user_email:
          anyOf:
          - type: string
          - type: 'null'
          title: User Email
        action:
          type: string
          title: Action
        resource_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Resource Type
        resource_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Resource Id
        status:
          type: string
          title: Status
        ip_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Ip Address
        user_agent:
          anyOf:
          - type: string
          - type: 'null'
          title: User Agent
        details:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Details
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - action
      - status
      - created_at
      title: AuditLogEntryResponse
    AuditLogListResponse:
      properties:
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
        entries:
          items:
            $ref: '#/components/schemas/AuditLogEntryResponse'
          type: array
          title: Entries
      type: object
      required:
      - total
      - limit
      - offset
      - entries
      title: AuditLogListResponse
    AuthProvider:
      type: string
      enum:
      - local
      - oauth2
      - saml
      - ldap
      title: AuthProvider
      description: "How a user's identity was established \u2014 see app/auth/sso/."
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RoleUpdateRequest:
      properties:
        role:
          $ref: '#/components/schemas/UserRole'
      type: object
      required:
      - role
      title: RoleUpdateRequest
    ScanJobCreateResponse:
      properties:
        scan_job_id:
          type: string
          format: uuid
          title: Scan Job Id
        repository_id:
          type: string
          format: uuid
          title: Repository Id
        status:
          $ref: '#/components/schemas/ScanJobStatus'
        priority:
          $ref: '#/components/schemas/ScanJobPriority'
        message:
          type: string
          title: Message
          default: Scan job queued
      type: object
      required:
      - scan_job_id
      - repository_id
      - status
      - priority
      title: ScanJobCreateResponse
...
    ScanJobPriority:
      type: string
      enum:
      - low
      - normal
      - high
      - critical
      title: ScanJobPriority
    ScanJobStatus:
      type: string
      enum:
      - queued
      - running
      - completed
      - failed
      - cancelled
      title: ScanJobStatus
...
    ScanJobSubmitRequest:
      properties:
        repo_url:
          type: string
          minLength: 1
          title: Repo Url
          description: GitHub, GitLab, or Bitbucket repository URL
        ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Ref
          description: "Branch, tag, or commit SHA \u2014 defaults to the provider's\
            \ default branch"
        priority:
          $ref: '#/components/schemas/ScanJobPriority'
          default: normal
        max_retries:
          type: integer
          maximum: 10.0
          minimum: 0.0
          title: Max Retries
          default: 2
        timeout_seconds:
          type: integer
          maximum: 7200.0
          minimum: 30.0
          title: Timeout Seconds
          default: 900
      type: object
      required:
      - repo_url
      title: ScanJobSubmitRequest
      description: "Body for `POST /scan/repository` \u2014 submit a repo URL to be\
        \ scanned."
...
    UserRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          format: email
          title: Email
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
          title: Is Active
        auth_provider:
          $ref: '#/components/schemas/AuthProvider'
      type: object
      required:
      - id
      - email
      - role
      - is_active
      - auth_provider
      title: UserRead
...
    UserRole:
      type: string
      enum:
      - admin
      - auditor
      - developer
      - security_engineer
      - read_only
      title: UserRole
      description: "The fixed RBAC role set \u2014 see app/auth/permissions.py for\
        \ what each\nrole can actually do (ROLE_PERMISSIONS). Not user-extensible:\
        \ adding a\nrole means adding an enum member + a migration, the same as any\
        \ other\nenum in this file, rather than a dynamic Role table \u2014 there\
        \ are 5 of\nthese by design, not 5000."
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WebhookAckResponse:
      properties:
        status:
          type: string
          title: Status
        message:
          type: string
          title: Message
        scan_job_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Scan Job Id
        repository_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Repository Id
      type: object
      required:
      - status
      - message
      title: WebhookAckResponse
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
...